# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "otelo",
    "description": "Query the logs, traces, and metrics that otelo keeps. Every path but the login and the logout needs the password of `otelo init`, as the cookie `otelo_password` or as `Authorization: Bearer <password>`.",
    "license": {
      "name": ""
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/services": {
      "get": {
        "summary": "The services that sent spans or logs in the range, the most requests\nfirst. A request is a span of the server kind with\n`http.request.method`. Each service has its requests and its logs over\nthe range and in buckets of one step, and the count of its spans.",
        "operationId": "list_services",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most services to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "step",
            "in": "query",
            "description": "The length of a bucket, such as `1m`. One that makes 60 buckets at\nmost when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Services"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{name}": {
      "get": {
        "summary": "One service: its requests and logs over the range and in buckets of one\nstep. A service without telemetry in the range has none of either.",
        "operationId": "get_service",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The name of the service",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "step",
            "in": "query",
            "description": "The length of a bucket, such as `1m`. One that makes 120 buckets at\nmost when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/spans": {
      "get": {
        "summary": "Spans, newest first.",
        "operationId": "list_spans",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most rows to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The records to keep, such as `http.route = \"/matches\" OR user.id = 7`.\nEvery record when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spans"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/spans/groups": {
      "get": {
        "summary": "The spans that the query keeps, grouped by the values of the `by` names,\nthe most time first. Each group has the count, the failures, the total\ntime, and the latency percentiles of its spans, and the name and the\nattributes of its newest span. The answer has the same numbers for all the\nspans, over the range and in buckets of one step.",
        "operationId": "list_span_groups",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most groups to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The spans to keep, such as `service = \"api\" kind = server`. Every span\nwhen missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "by",
            "in": "query",
            "description": "The names to group the spans by, separated by commas: attributes,\n`service`, `name`, or `resource.<key>`, such as\n`http.request.method,http.route`. All spans are one group when\nmissing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "step",
            "in": "query",
            "description": "The length of a bucket, such as `1m`. One that makes 120 buckets at\nmost when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpanGroups"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces": {
      "get": {
        "summary": "Traces with a span that the query keeps, by their root span, newest first.",
        "operationId": "list_traces",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most rows to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The records to keep, such as `http.route = \"/matches\" OR user.id = 7`.\nEvery record when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Traces"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{trace_id}": {
      "get": {
        "summary": "One trace: its spans by start time, and the logs that carry its ID.",
        "operationId": "get_trace",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "description": "The trace ID as 32 hex digits",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. The whole retention when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most spans, and the most logs, to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trace"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Attributes": {
        "type": "object",
        "description": "Attributes by key, in the order of their keys.",
        "additionalProperties": {
          "$ref": "#/components/schemas/AttributeValue"
        },
        "propertyNames": {
          "type": "string"
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Service": {
        "type": "object",
        "description": "One service in a range: its stats over time.",
        "required": [
          "service",
          "resource",
          "start_at",
          "end_at",
          "step_ns",
          "stats",
          "buckets"
        ],
        "properties": {
          "buckets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceBucket"
            },
            "description": "Every step of the range, oldest first."
          },
          "end_at": {
            "type": "string",
            "format": "date-time"
          },
          "resource": {
            "$ref": "#/components/schemas/Attributes",
            "description": "The attributes of the newest resource of the service. Empty when the\nrange has none of its telemetry."
          },
          "service": {
            "type": "string"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "description": "The range, after the retention capped it."
          },
          "stats": {
            "$ref": "#/components/schemas/ServiceStats"
          },
          "step_ns": {
            "type": "integer",
            "format": "int64",
            "description": "The length of a bucket in nanoseconds."
          }
        }
      },
      "ServiceBucket": {
        "type": "object",
        "description": "The requests and the logs of a service in one step.",
        "required": [
          "start_at",
          "requests",
          "logs",
          "error_logs"
        ],
        "properties": {
          "error_logs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "logs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "requests": {
            "$ref": "#/components/schemas/SpanStats"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "description": "The start of the step."
          }
        }
      },
      "ServiceStats": {
        "type": "object",
        "description": "The requests and the logs of a service in a range. A request is a span of\nthe server kind with `http.request.method`.",
        "required": [
          "requests",
          "spans",
          "logs",
          "error_logs"
        ],
        "properties": {
          "error_logs": {
            "type": "integer",
            "format": "int64",
            "description": "The logs of the error level and above.",
            "minimum": 0
          },
          "logs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "requests": {
            "$ref": "#/components/schemas/SpanStats"
          },
          "spans": {
            "type": "integer",
            "format": "int64",
            "description": "Every span of the service, of any kind.",
            "minimum": 0
          }
        }
      },
      "ServiceSummary": {
        "type": "object",
        "required": [
          "service",
          "resource",
          "stats",
          "buckets"
        ],
        "properties": {
          "buckets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceBucket"
            },
            "description": "Every step of the range, oldest first."
          },
          "resource": {
            "$ref": "#/components/schemas/Attributes",
            "description": "The attributes of the newest resource of the service that has any,\nsuch as `telemetry.sdk.language`."
          },
          "service": {
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/ServiceStats"
          }
        }
      },
      "Services": {
        "type": "object",
        "description": "The services that sent spans or logs in a range, the busiest first.",
        "required": [
          "start_at",
          "end_at",
          "step_ns",
          "services",
          "truncated"
        ],
        "properties": {
          "end_at": {
            "type": "string",
            "format": "date-time"
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceSummary"
            }
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "description": "The range, after the retention capped it."
          },
          "step_ns": {
            "type": "integer",
            "format": "int64",
            "description": "The length of a bucket in nanoseconds."
          },
          "truncated": {
            "type": "boolean",
            "description": "More services sent telemetry than the limit let through."
          }
        }
      },
...
      "SpanBucket": {
        "type": "object",
        "description": "The spans of one step.",
        "required": [
          "start_at",
          "spans"
        ],
        "properties": {
          "spans": {
            "$ref": "#/components/schemas/SpanStats"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "description": "The start of the step."
          }
        }
      },
...
      "SpanGroup": {
        "type": "object",
        "description": "The spans that share the values of the grouping fields.",
        "required": [
          "values",
          "name",
          "attributes",
          "spans"
        ],
        "properties": {
          "attributes": {
            "$ref": "#/components/schemas/Attributes",
            "description": "The attributes of its newest span."
          },
          "name": {
            "type": "string",
            "description": "The span name of its newest span."
          },
          "spans": {
            "$ref": "#/components/schemas/SpanStats"
          },
          "values": {
            "$ref": "#/components/schemas/Attributes",
            "description": "The values of the grouping fields, by the fields as a query writes\nthem, such as `http.route`. A field the spans lack is missing."
          }
        }
      },
      "SpanGroups": {
        "type": "object",
        "description": "The spans a query keeps in a range, grouped by the values of some fields,\nthe most time first.",
        "required": [
          "start_at",
          "end_at",
          "step_ns",
          "spans",
          "buckets",
          "groups",
          "truncated",
          "unindexed"
        ],
        "properties": {
          "buckets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpanBucket"
            },
            "description": "The spans of every step, oldest first."
          },
          "end_at": {
            "type": "string",
            "format": "date-time"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpanGroup"
            }
          },
          "spans": {
            "$ref": "#/components/schemas/SpanStats",
            "description": "Every span the query keeps."
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "description": "The range, after the retention capped it."
          },
          "step_ns": {
            "type": "integer",
            "format": "int64",
            "description": "The length of a bucket in nanoseconds."
          },
          "truncated": {
            "type": "boolean",
            "description": "More groups match than the limit let through, which kept the ones\nwith the most time."
          },
          "unindexed": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The attributes the query compares that have no index."
          }
        }
      },
...
      "Spans": {
        "type": "object",
        "description": "Spans, newest first.",
        "required": [
          "spans",
          "truncated",
          "unindexed"
        ],
        "properties": {
          "spans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraceSpan"
            }
          },
          "truncated": {
            "type": "boolean",
            "description": "More spans match than the limit let through."
          },
          "unindexed": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The attributes the query compares that have no index."
          }
        }
      },
...
      "Trace": {
        "type": "object",
        "description": "One trace: its spans by start time, and the logs that carry its ID.",
        "required": [
          "trace_id",
          "spans",
          "logs",
          "truncated"
        ],
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogLine"
            },
            "description": "Newest first."
          },
          "spans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraceSpan"
            }
          },
          "trace_id": {
            "type": "string"
          },
          "truncated": {
            "type": "boolean",
            "description": "The trace has more spans or logs than the limit let through."
          }
        }
      },
      "TraceSpan": {
        "type": "object",
        "required": [
          "trace_id",
          "span_id",
          "parent_span_id",
          "service",
          "name",
          "kind",
          "started_at",
          "duration_ns",
          "status",
          "attributes",
          "events",
          "resource"
        ],
        "properties": {
          "attributes": {
            "$ref": "#/components/schemas/Attributes"
          },
          "duration_ns": {
            "type": "integer",
            "format": "int64"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpanEvent"
            }
          },
          "kind": {
            "type": "integer",
            "format": "int32",
            "description": "The OpenTelemetry span kind."
          },
          "name": {
            "type": "string"
          },
          "parent_span_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "resource": {
            "$ref": "#/components/schemas/Attributes",
            "description": "The attributes of the resource that sent the span."
          },
          "service": {
            "type": "string"
          },
          "span_id": {
            "type": "string"
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The OpenTelemetry status code: 2 when the span failed."
          },
          "trace_id": {
            "type": "string"
          }
        }
      },
...
      "Traces": {
        "type": "object",
        "description": "Traces by their root span, newest first.",
        "required": [
          "traces",
          "truncated",
          "unindexed"
        ],
        "properties": {
          "traces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraceSummary"
            }
          },
          "truncated": {
            "type": "boolean",
            "description": "More traces match than the limit let through."
          },
          "unindexed": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The attributes the query compares that have no index."
          }
        }
      },
...
    }
  }
}